🚚 Move Plaquette-related enums to tqec.plaquette
