+++ src/tqec/plaquette/enums.py
