+++ src/tqec/plaquette/library/initialisation.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 import cirq
 
 from tqec.circuit.schedule import ScheduledCircuit
-from tqec.enums import PlaquetteOrientation
+from tqec.plaquette.enums import PlaquetteOrientation
 from tqec.plaquette.plaquette import Plaquette
 from tqec.plaquette.qubit import (
     PlaquetteQubits,
