+++ src/tqec/plaquette/library/measurement.py
@@ -2,7 +2,7 @@ import cirq
 
 from tqec.circuit.operations.operation import make_detector
 from tqec.circuit.schedule import ScheduledCircuit
-from tqec.enums import PlaquetteOrientation
+from tqec.plaquette.enums import PlaquetteOrientation
 from tqec.plaquette.plaquette import Plaquette
 from tqec.plaquette.qubit import (
     PlaquetteQubits,
